implement FIFO and fair scheduler for spark local mode
